Release 2, Latest Version added validation for register input fields changed design elements fixed bugs.

+++ app/models/Reading.java
@@ -11,7 +11,7 @@ public class Reading extends Model
 {
     public int code;
     public float temperature;
-    public double windSpeed;
+    public float windSpeed;
     public int pressure;
     public float windDirection;
